+++ ChangeLog.md
@@ -2,6 +2,15 @@
 Find here developer log of this package.    
 
 
+
+## 0.0.1 - 2020’-‘05’-‘27
+### Major change but still unstable
+- add package json builder
+- refactor all the code
+- add first attempt of documentation and sample
+- try of submodule with documentation
+- 
+
 ## 0.0.1 - 2020’-‘05’-‘27
 ### Start to create root of the project
 - add license